CSP-2379 Refactor: Rename GetProvidersFromLarsCode to ForLarsCode
