+++ src/SFA.DAS.Roatp.Api.UnitTests/Controllers/ExternalReadControllers/CoursesControllerTests/GetProvidersForLarsCodeV2Tests.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Mvc;
 using Moq;
 using NUnit.Framework;
 using SFA.DAS.Roatp.Api.Controllers.ExternalReadControllers;
-using SFA.DAS.Roatp.Application.Courses.Queries.GetProvidersFromLarsCode;
+using SFA.DAS.Roatp.Application.Courses.Queries.GetProvidersForLarsCode;
 using SFA.DAS.Roatp.Application.Mediatr.Responses;
 using SFA.DAS.Testing.AutoFixture;
 
@@ -19,7 +19,7 @@ namespace SFA.DAS.Roatp.Api.UnitTests.Controllers.ExternalReadControllers.Course
     {
         [Test, MoqAutoData]
         public async Task GetProvidersForLarsCodeV2_InvokesQueryHandler(
-            GetProvidersFromLarsCodeRequest request,
+            GetProvidersForLarsCodeRequest request,
             GetProvidersForLarsCodeQueryResult queryResult,
             [Frozen] Mock<IMediator> mediatorMock,
             [Greedy] CoursesController sut)
