+++ src/SFA.DAS.Roatp.Api/Models/V1/GetProvidersForLarsCodeResultModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using SFA.DAS.Roatp.Application.Courses.Queries.GetProvidersFromLarsCode;
+using SFA.DAS.Roatp.Application.Courses.Queries.GetProvidersForLarsCode;
 
 namespace SFA.DAS.Roatp.Api.Models.V1
 {
