+++ src/SFA.DAS.Roatp.Application.UnitTests/Courses/GetProvidersForLarsCode/GetProvidersForLarsCodeQueryValidatorTests.cs
@@ -4,7 +4,7 @@ using FluentValidation.TestHelper;
 using Moq;
 using NUnit.Framework;
 using SFA.DAS.Roatp.Application.Common;
-using SFA.DAS.Roatp.Application.Courses.Queries.GetProvidersFromLarsCode;
+using SFA.DAS.Roatp.Application.Courses.Queries.GetProvidersForLarsCode;
 using SFA.DAS.Roatp.Domain.Entities;
 using SFA.DAS.Roatp.Domain.Interfaces;
 using SFA.DAS.Roatp.Domain.Models;
@@ -27,7 +27,7 @@ public class GetProvidersForLarsCodeQueryValidatorTests
     [TestCase("1", true)]
     public async Task Validate_LarsCode(string larsCode, bool isValid)
     {
-        var request = new GetProvidersFromLarsCodeRequest();
+        var request = new GetProvidersForLarsCodeRequest();
         var validator = new GetProvidersForLarsCodeQueryValidator(_standardsReadRepo.Object);
 
         var result = await validator.TestValidateAsync(new GetProvidersForLarsCodeQuery(larsCode, request));
@@ -48,7 +48,7 @@ public class GetProvidersForLarsCodeQueryValidatorTests
     public async Task Validate_LatitudeRange(int latitude, bool isValid)
     {
         var larsCode = "1";
-        var request = new GetProvidersFromLarsCodeRequest
+        var request = new GetProvidersForLarsCodeRequest
         { Latitude = latitude, Longitude = 0, OrderBy = ProviderOrderBy.Distance };
         var validator = new GetProvidersForLarsCodeQueryValidator(_standardsReadRepo.Object);
 
@@ -78,7 +78,7 @@ public class GetProvidersForLarsCodeQueryValidatorTests
         string errorMessage)
     {
         var larsCode = "1";
-        var request = new GetProvidersFromLarsCodeRequest
+        var request = new GetProvidersForLarsCodeRequest
         { Latitude = latitude, Longitude = longitude, OrderBy = ProviderOrderBy.Distance };
         var validator = new GetProvidersForLarsCodeQueryValidator(_standardsReadRepo.Object);
 
@@ -96,7 +96,7 @@ public class GetProvidersForLarsCodeQueryValidatorTests
     public async Task Validate_OrderBy(ProviderOrderBy? orderBy, bool isValid)
     {
         var larsCode = "1";
-        var request = new GetProvidersFromLarsCodeRequest { Latitude = 80, Longitude = 0, OrderBy = orderBy };
+        var request = new GetProvidersForLarsCodeRequest { Latitude = 80, Longitude = 0, OrderBy = orderBy };
         var validator = new GetProvidersForLarsCodeQueryValidator(_standardsReadRepo.Object);
 
         var result = await validator.TestValidateAsync(new GetProvidersForLarsCodeQuery(larsCode, request));
@@ -118,7 +118,7 @@ public class GetProvidersForLarsCodeQueryValidatorTests
     public async Task Validate_Distance(int? distance, bool isValid)
     {
         var larsCode = "1";
-        var request = new GetProvidersFromLarsCodeRequest { Latitude = 80, Longitude = 0, OrderBy = ProviderOrderBy.Distance, Distance = distance };
+        var request = new GetProvidersForLarsCodeRequest { Latitude = 80, Longitude = 0, OrderBy = ProviderOrderBy.Distance, Distance = distance };
         var validator = new GetProvidersForLarsCodeQueryValidator(_standardsReadRepo.Object);
 
         var result = await validator.TestValidateAsync(new GetProvidersForLarsCodeQuery(larsCode, request));
