+++ src/SFA.DAS.Roatp.Application/Courses/Queries/GetProvidersForLarsCode/GetProvidersForLarsCodeQueryResult.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using SFA.DAS.Roatp.Domain.Models;
 
-namespace SFA.DAS.Roatp.Application.Courses.Queries.GetProvidersFromLarsCode
+namespace SFA.DAS.Roatp.Application.Courses.Queries.GetProvidersForLarsCode
 {
     public class GetProvidersForLarsCodeQueryResult
     {
